+++ mcu/attiny-hal/src/lib.rs
@@ -95,8 +95,6 @@ pub use adc::Adc;
 pub use port::Pins;
 
 #[cfg(feature = "device-selected")]
-pub mod wdt;
-#[cfg(feature = "device-selected")]
 pub use wdt::Wdt;
 
 #[cfg(feature = "device-selected")]
