+++ mcu/attiny-hal/src/periphals/mod.rs
@@ -5,3 +5,4 @@ pub mod eeprom;
 pub mod port;
 pub mod simple_pwm;
 pub mod spi;
+pub mod wdt;
